Fix printf format specifier mismatch for pid_t in process_table.c

+++ src/process_table.c
@@ -126,8 +126,8 @@ void process_table_add(struct process_table *pt, struct process *p) {
     if (locate_elem(pt->table[idx], &p->pid, offsetof(struct process, pid),
                     sizeof(pid_t)) == NULL) {
         if (add_elem(pt->table[idx], p) == NULL) {
-            fprintf(stderr, "Failed to add process with PID %d to the table\n",
-                    p->pid);
+            fprintf(stderr, "Failed to add process with PID %ld to the table\n",
+                    (long)p->pid);
             exit(EXIT_FAILURE);
         }
     }
